Checkif objects from turn order are not null or undefined
So far, problem is not exactly clear - I was not able to reproduce it.
During gaming session in some cases it fails with undefined exception,
when players do some acttions. Looks like it happens when there are some items in Turn order which
are hidden to GM layer. Tests are running without any exceptions or errors
even if object was not found. So, setting guardrails, just in case. If
problem reproduces in future - I'll investigate more.